Add curio.defer_cancellation and Task.cancel_no_wait
curio.defer_cancellation is an async context manager which defers any
cancellations until after leaving the 'async with' block. See
gh-96.

It *should* also defer timeout exceptions, but it doesn't yet.

cancel_no_wait is like cancel, but it doesn't wait for the task to
exit. I needed it to write a test for defer_cancellation, and was just
wanting it anyway for some user code I was working on...

I'm not sure whether it makes sense to have both cancel() and
cancel_no_wait() -- maybe cancel_no_wait() should be renamed to just
become cancel()? But for now I left both in.