Added python_requires to setup.py
Projects should ideally declare which Python versions they support. Curio 1.0 dropped Python 3.5 support but that was not reflected in the metadata in any way, which leads to syntax errors when doing `pip install curio`. The damage is already done for Python 3.5 but when 3.6 support is eventually dropped, this will help avoid problems then.